adiciona tokens de cores ao design system
